LlmClient - Meta LLaMA 3 (Scout) – 17B parameters, 16 experts (Mixture of Experts)
I switched not to use Ollama locally from the analysis computer, now I use LLaMA 3, which is LLaMA produced by meta. It's much faster than Ollama and I don't have to download it and run it locally because of the internet, but I use it with the api key I got by registering on GroqCloud.