ProtectedRoute
I created a ProtectedRoute so that no one can access the Admin, Student and Professor Dashboard without having the appropriate token with that role in it. Also, if a logged in user tries to go to the Login or Register page, they will not be able to, they will be redirected to their Dashboard.